+++ src/comm/enum/category.ts
@@ -1,4 +1,9 @@
-export enum Type {
-  EXPERIENCE = '경험',
-  FILE = '자료',
+export enum Category {
+  DEVELOPMENT = '개발',
+  ELECTRICAL_ELECTRONIC = '전기/전자',
+  MANUFACTURING = '생산/제조',
+  CONSTRUCTION_CIVIL = '건설/토목',
+  OFFICE_SERVICE = '사무/서비스',
+  MEDICAL = '의료',
+  ETC = '기타',
 }
