+++ src/comm/enum/type.ts
@@ -1,0 +1,4 @@
+export enum Type {
+  EXPERIENCE = '경험',
+  FILE = '자료',
+}
